# WebApiComplexFood/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ProjectStructure.Api",
    "version": "v1"
  },
  "paths": {
...
    "/carts/confirm/{buyerId}": {
      "patch": {
        "tags": [
          "ShoppingCart"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "buyerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ShoppingCartConfirmDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ShoppingCartConfirmDto"
            }
          }
        }
      }
    },
    "/carts/delete/{buyerId}": {
      "delete": {
        "tags": [
          "ShoppingCart"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "buyerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ShoppingCartConfirmDto": {
      "type": "object",
      "properties": {
        "couponCart": {
          "type": "string"
        }
      }
    },
...
  }
}